Introduce the notion of "acting as" a viewer
This notion is important so that even in a block like `omnisciently do ... end` or
`omnipotently do ... end`, the `Discretion.current_viewer` value is the actual logged-in viewer.

This also fixes a weird edge case where this was not previously possible:

```ruby
Discretion.omnisciently do
  ...
  Discretion.set_current_viewer(...)
end
```

In the previous implementation, the above code's call to `set_current_viewer()` would be a no-op
because Discretion resets the current_viewer to what it was before the omnisciently block.

Now, this code will work as intended.