Per-attribute privacy (#13)
Introducing more granularity with respect to attribute access on records.

If an AR is already successfully loaded (i.e. passed `can_see?(viewer)` check), we may want to
even further restrict access on a per-attribute level. An example is wanting to restrict
the `email` attribute of a `User` to all but that `User` themself (sic). That is, if another
`User` is logged in, we may never want to show the email of some other `User` to them, or we
may only want to show emails of a subset of `User`s (e.g. friends of the logged-in user).

Added specs to test this, and bumped version to 4.0.0.alpha (note not `pre` anymore).